[refactor] queryService 분리
